refactor: brokedown the monolithic aes key derivation file.
